app/requirements.txt
attrs==21.2.0
beautifulsoup4==4.9.3
bidict==0.21.2
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
cortex4py==2.1.0
d8s-dicts==0.6.0
d8s-hypothesis==0.6.0
d8s-lists==0.8.0
d8s-math==0.7.0
d8s-strings==0.5.0
d8s-uuids==0.6.0
dictdiffer==0.9.0
dnspython==1.16.0
emoji==1.4.2
eventlet==0.31.1
Flask==2.0.1
Flask-SocketIO==5.1.1
future==0.18.2
greenlet==1.1.1
hypothesis==6.15.0
idna==3.2
inflect==5.3.0
ioc-fanger==3.3.0
ioc-finder==6.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
more-itertools==8.8.0
mpmath==1.2.1
number-tools==0.1.0
pyparsing==2.4.7
python-engineio==4.2.1
python-magic==0.4.24
python-socketio==5.4.0
requests==2.26.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sympy==1.8
thehive4py==1.8.1
typing==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
